+++ sdk/storage/azure-storage-blob/azure/storage/blob/aio/download_async.py
@@ -48,7 +48,7 @@ class _AsyncChunkDownloader(_ChunkDownloader):
 
     async def process_chunk(self, chunk_start):
         chunk_start, chunk_end = self._calculate_range(chunk_start)
-        chunk_data = await self._download_chunk(chunk_start, chunk_end)
+        chunk_data = await self._download_chunk(chunk_start, chunk_end - 1)
         length = chunk_end - chunk_start
         if length > 0:
             await self._write_to_stream(chunk_data, chunk_start)
@@ -56,7 +56,7 @@ class _AsyncChunkDownloader(_ChunkDownloader):
 
     async def yield_chunk(self, chunk_start):
         chunk_start, chunk_end = self._calculate_range(chunk_start)
-        return await self._download_chunk(chunk_start, chunk_end)
+        return await self._download_chunk(chunk_start, chunk_end - 1)
 
     async def _update_progress(self, length):
         if self.progress_lock:
@@ -79,12 +79,12 @@ class _AsyncChunkDownloader(_ChunkDownloader):
 
         # No need to download the empty chunk from server if there's no data in the chunk to be downloaded.
         # Do optimize and create empty chunk locally if condition is met.
-        if self._do_optimize(download_range[0], download_range[1] - 1):
+        if self._do_optimize(download_range[0], download_range[1]):
             chunk_data = b"\x00" * self.chunk_size
         else:
             range_header, range_validation = validate_and_format_range_headers(
                 download_range[0],
-                download_range[1] - 1,
+                download_range[1],
                 check_content_md5=self.validate_content
             )
             try: