+++ sdk/storage/azure-storage-blob/tests/test_blob_service_properties.py
@@ -14,7 +14,7 @@ from azure.storage.blob import (
     BlobServiceClient,
     ContainerClient,
     BlobClient,
-    Logging,
+    BlobAnalyticsLogging,
     Metrics,
     CorsRule,
     RetentionPolicy,
@@ -43,7 +43,7 @@ class ServicePropertiesTest(StorageTestCase):
     def _assert_properties_default(self, prop):
         self.assertIsNotNone(prop)
 
-        self._assert_logging_equal(prop.logging, Logging())
+        self._assert_logging_equal(prop.logging, BlobAnalyticsLogging())
         self._assert_metrics_equal(prop.hour_metrics, Metrics())
         self._assert_metrics_equal(prop.minute_metrics, Metrics())
         self._assert_cors_equal(prop.cors, list())
@@ -122,7 +122,7 @@ class ServicePropertiesTest(StorageTestCase):
 
         # Act
         resp = self.bsc.set_service_properties(
-            logging=Logging(),
+            analytics_logging=BlobAnalyticsLogging(),
             hour_metrics=Metrics(),
             minute_metrics=Metrics(),
             cors=list(),
@@ -310,10 +310,10 @@ class ServicePropertiesTest(StorageTestCase):
     @record
     def test_set_logging(self):
         # Arrange
-        logging = Logging(read=True, write=True, delete=True, retention_policy=RetentionPolicy(enabled=True, days=5))
+        logging = BlobAnalyticsLogging(read=True, write=True, delete=True, retention_policy=RetentionPolicy(enabled=True, days=5))
 
         # Act
-        self.bsc.set_service_properties(logging=logging)
+        self.bsc.set_service_properties(analytics_logging=logging)
 
         # Assert
         received_props = self.bsc.get_service_properties()