+++ sdk/storage/azure-storage-blob/HISTORY.md
@@ -21,6 +21,7 @@
   - `encoding`
 - Removed types that were accidentally exposed from two modules. Only `BlobServiceClient`, `ContainerClient`, 
 `BlobClient` and `LeaseClient` should be imported from azure.storage.blob.aio
+- `Logging` has been renamed to `BlobAnalyticsLogging`.
 
 **New features**